draft xxx

+++ hospitalapi/src/main/java/org/ironhack/hospitalapi/repository/EmployeeRepository.java
@@ -1,6 +1,7 @@
 package org.ironhack.hospitalapi.repository;
 
+import org.ironhack.hospitalapi.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 }
